requirements.txt
fastapi
uvicorn
sqlmodel
python-jose[cryptography]
passlib[argon2]
python-multipart
requests
python-dotenv
resend
